Move payment provider implementations under domain
