+++ src/main/java/com/yi/online_bookshop/domain/service/paymentProvider/PaymentProviderInterface.java
@@ -1,4 +1,4 @@
-package com.yi.online_bookshop.domain.service;
+package com.yi.online_bookshop.domain.service.paymentProvider;
 
 public interface PaymentProviderInterface {
